requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
openai
faiss-cpu
tiktoken
python-dotenv
pdfplumber
python-docx